Added calib_file check

+++ run.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
+file_path="calibration_file/ur3e_calibration.yaml"
 # TODO: check whether this folder exists
-mkdir calibration_file
+if [ -e "$file_path" ]; then
+    echo "File exists in the folder."
+else
+    echo "Performing calibration"
+    mkdir calibration_file
 
-# Calibrate the robot
-docker run --rm --name "ur3e_calibration" \
+    # Calibrate the robot
+    docker run --rm --name "ur3e_calibration" \
     --privileged \
     --network "host" \
     --volume "$pwd/calibration_file":/calibration_file \
@@ -12,6 +17,7 @@ docker run --rm --name "ur3e_calibration" \
     bash -c "source /opt/ros/noetic/setup.bash && source /ur_ws/devel/setup.bash && \
              timeout 20 roslaunch ur_calibration calibration_correction.launch \
              robot_ip:=150.22.0.41 target_filename:='/calibration_file/ur3e_calibration.yaml'"
+fi
 
 # Load calibration file and start the robot
 docker run -d --name "ur3e_controller" \
@@ -22,6 +28,6 @@ docker run -d --name "ur3e_controller" \
     robotic_base:latest \
     bash -c "source /opt/ros/noetic/setup.bash && source /ur_ws/devel/setup.bash && \
              roslaunch ur_robot_driver ur3e_bringup.launch robot_ip:=150.22.0.41 \
-             kinematics_config:=/calibration_file/ur3e_calibration.yaml'"
+             kinematics_config:='/calibration_file/ur3e_calibration.yaml'"
